requirements.txt
accelerate==1.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
diffusers
filelock==3.13.1
fsspec==2024.6.1
furl==2.1.3
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
orderedmultidict==1.0.1
packaging==24.2
pillow==11.0.0
psutil==7.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
six==1.17.0
sympy==1.13.1
tabulate==0.9.0
tokenizers==0.21.0
tqdm==4.67.1
transformers
typing_extensions==4.12.2
urllib3==2.3.0
wget==3.2
zipp==3.21.0
